SPARK -1968 Replaced DummyTrustManager with SparkTrustManager in DummySSLSocketFactory. Moved DummySocketFactory into Certificates package to keep all TLS/certificates classes in one place and renamed it to SparkSSLSocketFactory (since it's no longer Dummy :) )
